+++ sifli-hal/Cargo.toml
@@ -35,7 +35,6 @@ cfg-if = { version = "1", features = ["core"] }
 futures-util = { version = "0.3.31", default-features = false }
 nb = "1.1.0"
 musb = { version = "0.3.0", features = ["embassy-usb-driver-impl", "prebuild"] }
-rwbt = { version = "0.1.0", features = ["sifli", "prebuild"] }
 
 # TODO: multicore critical-section impl
 cortex-m = { version = "0.7.7" }
@@ -56,7 +55,6 @@ portable-atomic = { version = "1.11.1", features = ["require-cas"] }
 display-driver = { version = "0.1.0" }
 bitfield-struct = "0.12.1"
 
-bt-hci = { version = "0.8", optional = true }
 rand_core = { version = "0.6" }
 
 [build-dependencies]
@@ -100,15 +98,6 @@ log = ["dep:log"]
 unstable-pac = []
 
 sf32lb52x = ["sifli-pac/sf32lb52x", "musb/builtin-sf32lb52x"]
-sf32lb52x-lcpu = ["sf32lb52x", "bt-hci"]
-
-## Enable EDR LO 3GHz calibration (VCO3G + OSLO). Only needed for BR/EDR (classic Bluetooth).
-## BLE-only applications can skip this for faster boot.
-edr-cal = ["sf32lb52x-lcpu"]
-
-## Enable bt-hci transport layer for IPC HCI communication
-bt-hci = ["dep:bt-hci", "sf32lb52x"]
-
 time-driver-atim1 = ["_time-driver-atim"]
 time-driver-gptim1 = ["_time-driver-gptim"]
 time-driver-gptim2 = ["_time-driver-gptim"]
